tendermint light client now compiles
