+++ light-clients/ics07-tendermint/Cargo.toml
@@ -1,0 +1,61 @@
+[package]
+name = "ics07-tendermint"
+version = "0.1.0"
+edition = "2021"
+
+[package.metadata.docs.rs]
+all-features = true
+
+[features]
+default = ["std"]
+std = []
+
+[dependencies]
+ibc = { git = "https://github.com/ComposableFi/ibc-rs", rev = "a3b905f1429734bd3e1fc65ddb9d91aa40ee8250", default-features = false }
+ibc-proto = { git = "https://github.com/ComposableFi/ibc-rs", rev = "a3b905f1429734bd3e1fc65ddb9d91aa40ee8250", default-features = false }
+
+borsh = { version = "0.9.3", default-features = false }
+ics23 = { git = "https://github.com/composablefi/ics23", rev = "b500a5c6068eb53c83c4c6c13bd9d8c25e0bf927", default-features = false }
+time = { version = "0.3", default-features = false }
+serde_derive = { version = "1.0.104", default-features = false }
+serde = { version = "1.0", default-features = false }
+serde_json = { version = "1", default-features = false }
+tracing = { version = "0.1.34", default-features = false }
+prost = { version = "0.10", default-features = false }
+prost-types = { version = "0.10", default-features = false }
+bytes = { version = "1.1.0", default-features = false }
+safe-regex = { version = "0.2.5", default-features = false }
+subtle-encoding = { version = "0.5", default-features = false }
+sha2 = { version = "0.10.2", default-features = false }
+flex-error = { version = "0.4.4", default-features = false }
+num-traits = { version = "0.2.15", default-features = false }
+derive_more = { version = "0.99.17", default-features = false, features = ["from", "into", "display"] }
+uint = { version = "0.9", default-features = false }
+primitive-types = { version = "0.11.1", default-features = false, features = ["serde_no_std"] }
+
+[dependencies.tendermint]
+git = "https://github.com/composableFi/tendermint-rs"
+rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8"
+default-features = false
+
+[dependencies.tendermint-proto]
+git = "https://github.com/composableFi/tendermint-rs"
+rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8"
+default-features = false
+
+[dependencies.tendermint-light-client-verifier]
+git = "https://github.com/composableFi/tendermint-rs"
+rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8"
+default-features = false
+
+
+[dev-dependencies]
+ibc = { git = "https://github.com/ComposableFi/ibc-rs", rev = "a3b905f1429734bd3e1fc65ddb9d91aa40ee8250", features = ["mocks"] }
+ibc-derive = { git = "https://github.com/ComposableFi/ibc-rs", rev = "a3b905f1429734bd3e1fc65ddb9d91aa40ee8250" }
+env_logger = "0.9.0"
+tracing-subscriber = { version = "0.3.11", features = ["fmt", "env-filter", "json"]}
+test-log = { version = "0.2.10", features = ["trace"] }
+modelator = "0.4.2"
+sha2 = { version = "0.10.2" }
+tendermint-rpc = { git = "https://github.com/composableFi/tendermint-rs", rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8", features = ["http-client", "websocket-client"] }
+tendermint-testgen = { git = "https://github.com/composableFi/tendermint-rs", rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8" } # Needed for generating (synthetic) light blocks.
