+++ light-clients/ics07-tendermint/src/mock/query/serialization/client_state.json
