+++ light-clients/ics07-tendermint/src/mock/query/serialization/consensus_state.json
