+++ light-clients/ics07-tendermint/src/query.rs
@@ -2,20 +2,22 @@
 
 use tendermint_rpc::abci::transaction::Hash;
 
-use ibc::core::ics02_client::client_consensus::QueryClientEventRequest;
-use ibc::core::ics04_channel::channel::QueryPacketEventDataRequest;
+use ibc::core::{
+	ics02_client::client_consensus::QueryClientEventRequest,
+	ics04_channel::channel::QueryPacketEventDataRequest,
+};
 
 /// Used for queries and not yet standardized in channel's query.proto
 #[derive(Clone, Debug)]
 pub enum QueryTxRequest {
-    Packet(QueryPacketEventDataRequest),
-    Client(QueryClientEventRequest),
-    Transaction(QueryTxHash),
+	Packet(QueryPacketEventDataRequest),
+	Client(QueryClientEventRequest),
+	Transaction(QueryTxHash),
 }
 
 #[derive(Clone, Debug)]
 pub enum QueryBlockRequest {
-    Packet(QueryPacketEventDataRequest),
+	Packet(QueryPacketEventDataRequest),
 }
 
 #[derive(Clone, Debug)]
