--- light-clients/ics07_tendermint/Cargo.toml
@@ -1,120 +1,0 @@
-[package]
-name = "clients"
-version = "0.1.0"
-edition = "2021"
-
-[package.metadata.docs.rs]
-all-features = true
-
-[features]
-default = ["std"]
-std = [
-    "flex-error/std",
-    "flex-error/eyre_tracer",
-    "ibc-proto/std",
-    "ics23/std",
-    "clock",
-    "beefy-client/std",
-    "sp-runtime/std",
-    "sp-core/std",
-    "codec/std",
-    "pallet-mmr-primitives/std",
-    "beefy-client-primitives/std",
-    "sp-trie/std",
-    "sp-io/std",
-    "sp-std/std",
-    "ibc/std"
-]
-clock = ["tendermint/clock", "time/std", "ibc/clock"]
-
-
-# This feature grants access to development-time mocking libraries, such as `MockContext` or `MockHeader`.
-# Depends on the `testgen` suite for generating Tendermint light blocks.
-mocks = ["clock", "std", "sha3", "ripemd", "ics11_beefy", "ibc/mocks"]
-ics11_beefy = ["sp-io", "sp-core", "sp-std", "beefy-primitives", "sp-runtime", "sp-trie", "pallet-mmr-primitives", "beefy-client", "beefy-client-primitives" ]
-
-[dependencies]
-ibc = { path = "../modules" }
-# Proto definitions for all IBC-related interfaces, e.g., connections or channels.
-borsh = { version = "0.9.3", default-features = false }
-ibc-proto = { version = "0.18.0", path = "../proto", default-features = false }
-ics23 = { git = "https://github.com/composablefi/ics23", rev = "b500a5c6068eb53c83c4c6c13bd9d8c25e0bf927", default-features = false }
-time = { version = "0.3", default-features = false }
-serde_derive = { version = "1.0.104", default-features = false }
-serde = { version = "1.0", default-features = false }
-serde_json = { version = "1", default-features = false }
-tracing = { version = "0.1.34", default-features = false }
-prost = { version = "0.10", default-features = false }
-prost-types = { version = "0.10", default-features = false }
-bytes = { version = "1.1.0", default-features = false }
-safe-regex = { version = "0.2.5", default-features = false }
-subtle-encoding = { version = "0.5", default-features = false }
-sha2 = { version = "0.10.2", default-features = false }
-flex-error = { version = "0.4.4", default-features = false }
-num-traits = { version = "0.2.15", default-features = false }
-derive_more = { version = "0.99.17", default-features = false, features = ["from", "into", "display"] }
-uint = { version = "0.9", default-features = false }
-beefy-client = { package = "beefy-light-client",  git = "https://github.com/ComposableFi/beefy-rs", rev = "cb8cadc8d45bc444367002c77cbd395eff8a741c", default-features = false, optional = true }
-beefy-client-primitives = { package = "primitives", git = "https://github.com/ComposableFi/beefy-rs", rev = "cb8cadc8d45bc444367002c77cbd395eff8a741c", default-features = false, optional = true }
-pallet-mmr-primitives = { package = "sp-mmr-primitives", git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24", default-features = false, optional = true }
-beefy-primitives = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24", default-features = false, optional = true }
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false }
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24", default-features = false, optional = true }
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24", default-features = false, optional = true}
-sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24", default-features = false, optional = true }
-sp-trie = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24", default-features = false, optional = true }
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24", default-features = false, optional = true }
-sha3 = { version = "0.10.1", optional = true }
-ripemd = { version = "0.1.1", optional = true }
-
-primitive-types = { version = "0.11.1", default-features = false, features = ["serde_no_std"] }
-derive = { path = "../derive", package = "ibc-derive" }
-
-[dependencies.tendermint]
-git = "https://github.com/composableFi/tendermint-rs"
-rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8"
-default-features = false
-
-[dependencies.tendermint-proto]
-git = "https://github.com/composableFi/tendermint-rs"
-rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8"
-default-features = false
-
-[dependencies.tendermint-light-client-verifier]
-git = "https://github.com/composableFi/tendermint-rs"
-rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8"
-default-features = false
-
-
-[dev-dependencies]
-env_logger = "0.9.0"
-tracing-subscriber = { version = "0.3.11", features = ["fmt", "env-filter", "json"]}
-test-log = { version = "0.2.10", features = ["trace"] }
-modelator = "0.4.2"
-sha2 = { version = "0.10.2" }
-tendermint-rpc = { git = "https://github.com/composableFi/tendermint-rs", rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8", features = ["http-client", "websocket-client"] }
-tendermint-testgen = { git = "https://github.com/composableFi/tendermint-rs", rev = "5a74e0f8da4d3dab83cc04b5f1363b018cf3d9e8" } # Needed for generating (synthetic) light blocks.
-# Beefy Light Client testing dependencies
-sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24"}
-subxt = { git = "https://github.com/paritytech/subxt", rev = "ec23283d75e4b3b894294e351fd7ffa2b4431201" }
-tokio = { version = "1.17.0", features = ["full"] }
-serde_json = "1.0.74"
-beefy-client = { package = "beefy-light-client",  git = "https://github.com/ComposableFi/beefy-rs", rev = "cb8cadc8d45bc444367002c77cbd395eff8a741c" }
-beefy-queries = { package = "beefy-prover", git = "https://github.com/ComposableFi/beefy-rs", rev = "cb8cadc8d45bc444367002c77cbd395eff8a741c" }
-sha3 = { version = "0.10.1" }
-ripemd = { version = "0.1.1" }
-beefy-client-primitives = { package = "primitives", git = "https://github.com/ComposableFi/beefy-rs", rev = "cb8cadc8d45bc444367002c77cbd395eff8a741c" }
-beefy-primitives = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24" }
-pallet-mmr-primitives = { package = "sp-mmr-primitives", git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24" }
-codec = { package = "parity-scale-codec", version = "3.0.0"}
-sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24" }
-sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24" }
-sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24" }
-sp-trie = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24" }
-frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.24" }
-ibc = { path = "../modules", features = ['mocks'] }
-
-[[test]]
-name = "mbt"
-path = "tests/mbt.rs"
-required-features = ["mocks"]
